important refactoring
The existing models were refactored, features suitable for each model were selected, and the data initialization code was rewritten. In the new version, the model can generally obtain more than 90% prediction accuracy for correct samples when there is a one-to-one ratio of benign and malignant samples in the training set.