Version rollback
In the latest version (version 35, df0c847), an unknown error caused the project to crash. After multiple checks and verifications, no obvious source of the error could be found. This error caused the 0.5 version of IDS under development and even the 0.4 version that had been developed to fail to work at the same time. In the end, I rolled back the version of this project based on a backup on another machine. The rolled back version should be basically the same as version 33 (df6572d).

However, compared to version 33, this version has more detailed code on how to find applicable features and the feature sets that have been collected.